Conflicts:
	Android.mk
	hostapd/Android.mk
	src/drivers/driver.h
	src/drivers/driver_nl80211.c
	src/drivers/driver_test.c
	src/p2p/p2p.c
	src/p2p/p2p.h
	src/p2p/p2p_group.c
	src/p2p/p2p_i.h
	src/p2p/p2p_pd.c
	src/utils/common.c
	wpa_supplicant/Android.mk
	wpa_supplicant/config.c
	wpa_supplicant/config.h
	wpa_supplicant/ctrl_iface.c
	wpa_supplicant/dbus/dbus_new_handlers_p2p.c
	wpa_supplicant/driver_i.h
	wpa_supplicant/events.c
	wpa_supplicant/p2p_supplicant.c
	wpa_supplicant/p2p_supplicant.h
	wpa_supplicant/wpa_cli.c
	wpa_supplicant/wpa_supplicant.c
	wpa_supplicant/wpa_supplicant.conf
	wpa_supplicant/wpa_supplicant_conf.mk
	wpa_supplicant/wpa_supplicant_i.h

+++ src/eap_server/eap_server_ttls.c
@@ -2,14 +2,8 @@
  * hostapd / EAP-TTLS (RFC 5281)
  * Copyright (c) 2004-2011, Jouni Malinen <j@w1.fi>
  *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 2 as
- * published by the Free Software Foundation.
- *
- * Alternatively, this software may be distributed under the terms of BSD
- * license.
- *
- * See README and COPYING for more details.
+ * This software may be distributed under the terms of the BSD license.
+ * See README for more details.
  */
 
 #include "includes.h"
@@ -98,7 +92,8 @@ static u8 * eap_ttls_avp_hdr(u8 *avphdr, u32 avp_code, u32 vendor_id,
 	}
 
 	avp->avp_code = host_to_be32(avp_code);
-	avp->avp_length = host_to_be32((flags << 24) | (hdrlen + len));
+	avp->avp_length = host_to_be32(((u32) flags << 24) |
+				       ((u32) (hdrlen + len)));
 
 	return avphdr + hdrlen;
 }
